Fix picking/hover to return closest object.
Add `View.pickClosest(xp, yp, maxDistance)` function to select / return
a single object that is the closest to the mouse pointer position.
Search region is restricted based on the `pointerSize` property.
Earlier, the function simply returned/handled the first object in the
`selection` array, which was unpredictable.
This change affects the 'click' and 'hover' picking modes.
The 'select' picking mode (which uses a shift + click-drag a
rectangular region) continues to use the previous implementation,
which returns all objects within that rectangular region (selection
zone) in no particular order.